Updated stubs
I am going to remove these service stubs  and interfaces in next commit because the code in these files should be in a DAO layer - which I have created.

Note - I am still committing changes (that are required to these files as a result of changes made to DTOs) in case you guys want to just merge the DTO changes.